Feature: Full screen shortcut support
Tested with iPad ios 13 safari, Samsung Browser and Chrome on Android.
This supports full screen shortcuts when "add to desktop" or the equivalent is selected from within the web browser while on the komga page. This will trigger the web browser's full screen stand alone web app mode. It is not a full PWA implementation, so offline caching does not occur.